Add wrapper for class, decorators

+++ tests/valgrind_parser_test.py
@@ -6,7 +6,7 @@ class TestValgrindParser(object):
 
     def teardown_class(cls):
         import os
-        os.remove(cls.html_report_location)
+        if os.path.exists(cls.html_report_location): os.remove(cls.html_report_location)
 
     def test_json_helper(self):
         jsonPath = 'valgrind_parser/valgrind_parser/data/valgrind_regexes.json'
@@ -30,3 +30,27 @@ class TestValgrindParser(object):
         }
 
         dump_html_report(errors_dict=errors_dict, html_report_location=self.html_report_location)
+
+    def test_generate_valgrind_report_function(self):
+        from valgrind_parser import generate_valgrind_report
+        logPath = 'valgrind_parser/test_dir/valgrind_log.txt'
+        generate_valgrind_report(logPath)
+
+    def test_decorator_trycatch(self):
+        from valgrind_parser.valgrind_parser.utils.decorators import trycatch
+        @trycatch
+        def internal_test_function():
+            raise Exception
+
+        internal_test_function()
+
+    def test_decorator_singleton(self):
+        from valgrind_parser.valgrind_parser.utils.decorators import singleton
+        @singleton
+        class TestClass:
+            pass
+
+        object1 = TestClass()
+        object2 = TestClass()
+
+        assert object1 is object2, "Test class objects are different"
